chore(code): Rename storage -> local
- Align with idb.ts name
