+++ src/shared/theme/useTheme.ts
@@ -1,7 +1,7 @@
 import create from 'zustand';
 
 import { ThemeMode } from '../../types';
-import { getValue } from '../storage/storage';
+import { getValue } from '../storage/local';
 
 type ThemeState = {
   theme: ThemeMode;
